make a WebpubViewer component to be shared with /read routes
