+++ .gitignore
@@ -31,5 +31,4 @@ yalc.lock
 wallaby.js
 config/
 .vscode/
-lib/
 .vercel
