+++ src/pages/read/[bookUrl].tsx
@@ -1,978 +1,8 @@
 import React from "react";
-import reader from "utils/reader";
-import { useRouter } from "next/router";
-import useLibraryContext from "../../components/context/LibraryContext";
+import WebpubViewer from "../../components/WebpubViewer";
 
 const BookPage = () => {
-  const library = useLibraryContext();
-  const router = useRouter();
-  const { bookUrl } = router.query;
-  const bookManifestUrl = `${bookUrl}/META-INF/container.xml`;
-
-  const { catalogName } = library;
-
-  if (typeof window !== "undefined") {
-    reader(bookManifestUrl, catalogName);
-  }
-
-  return (
-    <>
-      <div id="viewer" />
-      <style global jsx>{`
-        :root {
-          -webkit-text-size-adjust: 100%;
-          zoom: reset;
-        }
-
-        body {
-          font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
-            Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji",
-            "Segoe UI Symbol";
-          font-size: 16px;
-          margin: 0 auto;
-        }
-
-        ul li,
-        ol li {
-          list-style: none;
-        }
-
-        button {
-          background: #fff;
-          border: 0;
-          color: #5b5852;
-          padding: 0.35rem 0.35rem 0.2rem;
-          margin: 0;
-          -webkit-appearance: none;
-        }
-
-        a {
-          color: #5b5852;
-          text-decoration: none;
-        }
-
-        .inactive {
-          opacity: 0;
-          z-index: -3000;
-        }
-
-        .active {
-          opacity: 1;
-          z-index: 3000;
-        }
-
-        #viewer {
-          padding: 0 env(safe-area-inset-right) 0 env(safe-area-inset-left);
-          box-sizing: border-box;
-        }
-
-        .svgIcon.use {
-          display: none;
-          height: 0;
-          position: absolute;
-          width: 0;
-        }
-
-        .controls-trigger {
-          position: fixed;
-          top: 0.5rem;
-          right: 0.5rem;
-          padding: env(safe-area-inset-top) 0 0;
-          z-index: 10;
-          background-color: #fff;
-          color: #5b5852;
-        }
-        .controls-trigger .icon {
-          display: inline-block;
-          color: #5b5852;
-          fill: #5b5852;
-          height: 1.5rem;
-          width: 1.5rem;
-          transition: transform 300ms ease-in-out;
-        }
-        .controls-trigger .inactive-icon {
-          transition: transform 300ms ease-in-out;
-          transform: rotate(180deg);
-        }
-
-        .controls {
-          color: #5b5852;
-          position: fixed;
-          right: 0;
-          top: 0;
-          width: 100%;
-          padding: env(safe-area-inset-top) 0 0;
-        }
-        .controls ul li,
-        .controls ol li {
-          display: inline-block;
-        }
-        .controls .links {
-          background-color: #fff;
-          height: inherit;
-          list-style-type: none;
-          margin: 0;
-          position: relative;
-          padding: 0.125rem;
-          text-align: center;
-        }
-        .controls .links .icon {
-          display: inline-block;
-          color: #5b5852;
-          fill: #5b5852;
-          height: 1.5rem;
-          width: 1.5rem;
-        }
-        .controls .links .icon.inactive-icon {
-          display: none;
-          height: 0;
-          width: 0;
-        }
-        .controls .links .setting-text {
-          display: none;
-          position: relative;
-          text-transform: uppercase;
-          top: -0.4rem;
-        }
-        @media screen and (min-width: 60rem) {
-          .controls .links .setting-text {
-            display: inline-block;
-            font-size: 0.9rem;
-            letter-spacing: 0.07rem;
-          }
-        }
-        .controls .links.top {
-          background-color: #fff;
-          border-bottom: 0.0625rem solid #dadada;
-          height: 2.9rem;
-          padding: 0.5rem 0 0;
-          display: flex;
-        }
-        .controls .links.top.inactive {
-          transition: opacity 100ms cubic-bezier(0.32, 0.76, 1, 1.01);
-        }
-        .controls .links.top.active {
-          transition: opacity 100ms cubic-bezier(0, 0.03, 0.63, 0.99);
-        }
-        .controls .links.top > li:nth-child(1) {
-          display: inline-block;
-          margin: 0 auto 0 0.625rem;
-        }
-        .controls .links.top > li:nth-child(1) a[rel="up"] {
-          cursor: pointer;
-          display: inline-block;
-          padding: 0.35rem 0.35rem 0.2rem;
-        }
-        .controls .links.top > li.uplink-wrapper {
-          margin-right: auto;
-        }
-        .controls .links.top > li:last-child {
-          margin-right: 0.625rem;
-          margin-right: 2.85rem;
-        }
-        .controls .links.bottom {
-          background-color: #f9f9f9;
-          border-top: 0.16rem solid #dadada;
-          bottom: 0;
-          height: 2.2rem;
-          padding-top: 0.625rem;
-          position: fixed;
-          width: 100%;
-        }
-        .controls .links.bottom.inactive {
-          height: 0;
-        }
-        @media (min-width: 30rem) {
-          .controls .links.bottom {
-            padding-top: 0;
-          }
-        }
-        .controls .links.bottom li {
-          border: 0;
-          float: none;
-          padding: 0;
-        }
-        @media screen and (min-width: 30rem) {
-          .controls .links.bottom li:nth-child(1) {
-            float: left;
-          }
-          .controls .links.bottom li:nth-child(3) {
-            float: right;
-          }
-        }
-        .controls .links.bottom li[aria-label="chapters"] {
-          margin: 0 20%;
-          position: relative;
-          top: -0.34rem;
-          text-align: center;
-        }
-        @media screen and (min-width: 30rem) {
-          .controls .links.bottom li[aria-label="chapters"] {
-            display: none;
-          }
-        }
-        @media screen and (min-width: 30rem) {
-          .controls .links.bottom {
-            background: #fff;
-          }
-        }
-        .controls .links.bottom .icon {
-          height: 1.5rem;
-          width: 1.5rem;
-          margin: 0 0.375rem;
-        }
-        @media screen and (min-width: 30rem) {
-          .controls .links.bottom .icon {
-            position: relative;
-            top: 0.375rem;
-          }
-        }
-        .controls .links.bottom .chapter-control {
-          display: none;
-        }
-        @media screen and (min-width: 30rem) {
-          .controls .links.bottom .chapter-control {
-            display: inline-block;
-          }
-        }
-
-        @-webkit-keyframes load {
-          0% {
-            transform: translate(-2.5rem, -2.5rem) rotate(0deg);
-          }
-          100% {
-            transform: translate(-2.5rem, -2.5rem) rotate(360deg);
-          }
-        }
-
-        @keyframes load {
-          0% {
-            transform: translate(-2.5rem, -2.5rem) rotate(0deg);
-          }
-          100% {
-            transform: translate(-2.5rem, -2.5rem) rotate(360deg);
-          }
-        }
-
-        .loading.is-loading .icon {
-          animation: load 1s ease-in-out infinite;
-        }
-
-        .loading {
-          position: fixed;
-          width: 100%;
-          height: 100vh;
-          z-index: 10;
-          background-color: rgba(255, 255, 255, 0.9);
-          color: #5b5852;
-          -webkit-user-select: none;
-          -moz-user-select: none;
-          -ms-user-select: none;
-          user-select: none;
-          cursor: default;
-        }
-        .loading .icon {
-          display: block;
-          position: absolute;
-          top: 50%;
-          left: 50%;
-          width: 5rem;
-          height: 5rem;
-          transform: translate(-2.5rem, -2.5rem);
-          fill: #5b5852;
-        }
-
-        .error {
-          z-index: 20;
-          background-color: rgba(255, 255, 255, 0.875);
-          color: #d0343a;
-          height: 100%;
-          padding-top: 40vh;
-          position: fixed;
-          text-align: center;
-          width: 100%;
-          -webkit-user-select: none;
-          -moz-user-select: none;
-          -ms-user-select: none;
-          user-select: none;
-          cursor: default;
-        }
-        .error .icon {
-          width: 3rem;
-          height: 3rem;
-          fill: #d0343a;
-        }
-        .error span {
-          display: block;
-          margin-bottom: 0.75rem;
-        }
-        .error button {
-          color: #d0343a;
-          border: 0.125rem solid #d0343a;
-          border-radius: 0.3125rem;
-          font-size: 1rem;
-          font-weight: 700;
-          padding: 1.4rem;
-        }
-        .error button + button {
-          margin-left: 1rem;
-        }
-
-        .controls-view {
-          font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
-            Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji",
-            "Segoe UI Symbol";
-          font-size: 16px;
-          height: calc(100% - 2rem);
-          left: 0;
-          position: fixed;
-          top: 3rem;
-          width: 100%;
-          -webkit-overflow-scrolling: touch;
-          text-align: left;
-        }
-        .controls-view .disabled {
-          background-color: #a7a7a7;
-        }
-        .controls-view.inactive {
-          transition: opacity 300ms cubic-bezier(0, 0.03, 0.63, 0.99);
-          height: 0;
-        }
-        .controls-view.active {
-          transition: opacity 300ms cubic-bezier(0, 0.03, 0.63, 0.99);
-        }
-        .controls-view > ul {
-          padding-bottom: 1rem;
-        }
-
-        .info {
-          color: #5b5852;
-          margin: 0;
-          padding: 0 1.5rem;
-          overflow: hidden;
-          text-align: center;
-          text-overflow: ellipsis;
-          white-space: nowrap;
-          -webkit-user-select: none;
-          -moz-user-select: none;
-          -ms-user-select: none;
-          user-select: none;
-          cursor: default;
-        }
-        .info.top {
-          line-height: 3;
-          padding-top: env(safe-area-inset-top);
-        }
-        .info.bottom {
-          line-height: 2;
-          padding-bottom: env(safe-area-inset-bottom);
-        }
-        .info .chapter-position,
-        .info .chapter-title {
-          font-size: 0.85rem;
-          font-variant-numeric: lining-nums tabular-nums;
-        }
-
-        iframe {
-          opacity: 0;
-        }
-        iframe.left-hover {
-          background: linear-gradient(
-            to right,
-            rgba(218, 218, 218, 0.9),
-            rgba(218, 218, 218, 0.9) 28%,
-            #fff 30%,
-            #fff
-          );
-        }
-        iframe.right-hover {
-          background: linear-gradient(
-            to left,
-            rgba(218, 218, 218, 0.9),
-            rgba(218, 218, 218, 0.9) 28%,
-            #fff 30%,
-            #fff
-          );
-        }
-
-        .flip-page-container {
-          height: 100px;
-          margin: auto;
-          display: grid;
-          align-content: center;
-        }
-        @media screen and (max-width: 40rem) {
-          .flip-page-container {
-            width: 3rem;
-          }
-        }
-        .flip-page-container.hidden {
-          display: none;
-        }
-
-        .flip-page-btn {
-          display: inline-block;
-          border: 1px solid #d5d5d5;
-          background: #fefefe;
-          fill: #5b5852;
-          width: 50px;
-          height: 50px;
-          border-radius: 25px;
-        }
-        @media screen and (max-width: 40rem) {
-          .flip-page-btn {
-            display: none;
-          }
-        }
-        .flip-page-btn.prev {
-          margin: auto 4em 0 2em;
-        }
-        .flip-page-btn.next {
-          margin: auto 2em 0 4em;
-        }
-
-        .flip-page-container:hover .flip-page-btn {
-          background: #4d4d4d;
-          fill: #fefefe;
-          border: 1px solid #494949;
-          stroke: #fefefe;
-        }
-
-        .flip-page-icon {
-          height: 100%;
-        }
-        .flip-page-icon.next {
-          transform: rotate(-90deg);
-        }
-        .flip-page-icon.prev {
-          transform: rotate(90deg);
-        }
-
-        .page-container {
-          display: flex;
-        }
-
-        .scrolling-suggestion {
-          position: absolute;
-          left: -10000px;
-          top: auto;
-          width: 0.0625rem;
-          height: 0.0625rem;
-          overflow: hidden;
-          background-color: #fff;
-          margin: 0.3125rem;
-          padding: 0.3125rem 0.625rem;
-          z-index: 4000;
-        }
-        .scrolling-suggestion:focus {
-          left: auto;
-          width: auto;
-          height: auto;
-          overflow: visible;
-          top: 0.3125rem;
-        }
-
-        [data-viewer-theme="sepia"] {
-          background-color: #f6ecd9;
-        }
-        [data-viewer-theme="sepia"] button {
-          background: #f6ecd9;
-          color: #5b5852;
-        }
-        [data-viewer-theme="sepia"] a {
-          color: #5b5852;
-        }
-        [data-viewer-theme="sepia"] .controls-trigger {
-          background: #f6ecd9;
-          color: #5b5852;
-        }
-        [data-viewer-theme="sepia"] .controls-trigger .icon {
-          color: #5b5852;
-          fill: #5b5852;
-        }
-        [data-viewer-theme="sepia"] .controls {
-          color: #5b5852;
-        }
-        [data-viewer-theme="sepia"] .controls .links {
-          background-color: #f6ecd9;
-        }
-        [data-viewer-theme="sepia"] .controls .links .icon {
-          color: #5b5852;
-          fill: #5b5852;
-        }
-        [data-viewer-theme="sepia"] .controls.top {
-          background-color: #f6ecd9;
-          border-bottom: 0.0625rem solid #dadada;
-        }
-        [data-viewer-theme="sepia"] .controls.bottom {
-          background-color: #f4e7cf;
-          border-top: 0.16rem solid #dadada;
-        }
-        [data-viewer-theme="sepia"] .loading {
-          background-color: rgba(246, 236, 217, 0.9);
-        }
-        [data-viewer-theme="sepia"] .error {
-          background-color: rgba(246, 236, 217, 0.875);
-        }
-        [data-viewer-theme="sepia"] .error button {
-          background-color: #fff;
-          color: #d0343a;
-          border: 0.125rem solid #d0343a;
-        }
-        [data-viewer-theme="sepia"] .scrolling-suggestion {
-          background-color: #f6ecd9;
-        }
-        [data-viewer-theme="sepia"] .info {
-          color: #5b5852;
-        }
-        [data-viewer-theme="sepia"] iframe.left-hover {
-          background: linear-gradient(
-            to right,
-            #ecd7b0,
-            #ecd7b0 28%,
-            #f6ecd9 30%,
-            #f6ecd9
-          );
-        }
-        [data-viewer-theme="sepia"] iframe.right-hover {
-          background: linear-gradient(
-            to left,
-            #ecd7b0,
-            #ecd7b0 28%,
-            #f6ecd9 30%,
-            #f6ecd9
-          );
-        }
-
-        [data-viewer-theme="night"] {
-          background-color: #111;
-          color: #fff;
-        }
-        [data-viewer-theme="night"] button {
-          background: #111;
-          color: #dadada;
-        }
-        [data-viewer-theme="night"] a {
-          color: #dadada;
-        }
-        [data-viewer-theme="night"] .controls-trigger {
-          background: #111;
-          color: #dadada;
-        }
-        [data-viewer-theme="night"] .controls-trigger .icon {
-          color: #dadada;
-          fill: #dadada;
-        }
-        [data-viewer-theme="night"] .controls {
-          color: #dadada;
-        }
-        [data-viewer-theme="night"] .controls .links {
-          background-color: #111;
-        }
-        [data-viewer-theme="night"] .controls .links .icon {
-          color: #dadada;
-          fill: #dadada;
-        }
-        [data-viewer-theme="night"] .controls.top {
-          background-color: #111;
-          border-bottom: 0.0625rem solid #5b5852;
-        }
-        [data-viewer-theme="night"] .controls.bottom {
-          background-color: #171717;
-          border-top: 0.16rem solid #5b5852;
-        }
-        [data-viewer-theme="night"] .loading {
-          background-color: #2b2b2b;
-          color: #dadada;
-        }
-        [data-viewer-theme="night"] .loading .icon {
-          fill: #dadada;
-        }
-        [data-viewer-theme="night"] .error {
-          background-color: rgba(17, 17, 17, 0.875);
-        }
-        [data-viewer-theme="night"] .error button {
-          background-color: #2b2b2b;
-          color: #d0343a;
-          border: 0.125rem solid #d0343a;
-        }
-        [data-viewer-theme="night"] .scrolling-suggestion {
-          background-color: #111;
-        }
-        [data-viewer-theme="night"] .info {
-          color: #dadada;
-        }
-        [data-viewer-theme="night"] iframe.left-hover {
-          background: linear-gradient(
-            to right,
-            #2b2b2b,
-            #2b2b2b 28%,
-            #111 30%,
-            #111
-          );
-        }
-        [data-viewer-theme="night"] iframe.right-hover {
-          background: linear-gradient(
-            to left,
-            #2b2b2b,
-            #2b2b2b 28%,
-            #111 30%,
-            #111
-          );
-        }
-
-        .contents-view {
-          background-color: #fff;
-          overflow: scroll;
-          top: 3.2rem;
-          -webkit-user-select: none;
-          -moz-user-select: none;
-          -ms-user-select: none;
-          user-select: none;
-        }
-        .contents-view ol li {
-          margin-top: 0;
-          width: 88%;
-        }
-        .contents-view ol li a {
-          color: #5b5852;
-          border-bottom: 1px solid #cccccc;
-          display: block;
-          padding: 1rem;
-          text-decoration: none;
-          width: 100%;
-        }
-        @media screen and (min-width: 60rem) {
-          .contents-view ol li a:hover {
-            background: #5b5852;
-            color: #fff;
-          }
-        }
-        .contents-view ol li a.active {
-          background: #dadada;
-          color: #111;
-        }
-        .contents-view ol li a.active:hover {
-          background: #111;
-          color: #dadada;
-        }
-        .contents-view ol li span {
-          color: #5b5852;
-          border-bottom: 1px solid #cccccc;
-          display: block;
-          padding: 1rem;
-          text-decoration: none;
-          width: 100%;
-        }
-
-        [data-viewer-theme="sepia"] .contents-view {
-          background-color: #f6ecd9;
-        }
-        [data-viewer-theme="sepia"] .contents-view ol li a {
-          color: #5b5852;
-          border-bottom: 1px solid #e2c387;
-        }
-        @media screen and (min-width: 60rem) {
-          [data-viewer-theme="sepia"] .contents-view ol li a:hover {
-            background: #5b5852;
-            color: #f6ecd9;
-          }
-        }
-        [data-viewer-theme="sepia"] .contents-view ol li a.active {
-          background: #dadada;
-          color: #111;
-        }
-        [data-viewer-theme="sepia"] .contents-view ol li a.active:hover {
-          background: #111;
-          color: #dadada;
-        }
-        [data-viewer-theme="sepia"] .contents-view ol li span {
-          color: #5b5852;
-          border-bottom: 1px solid #e2c387;
-        }
-
-        [data-viewer-theme="night"] .contents-view {
-          background-color: #111;
-        }
-        [data-viewer-theme="night"] .contents-view ol li a {
-          color: #fff;
-          border-bottom: 1px solid #444444;
-        }
-        @media screen and (min-width: 60rem) {
-          [data-viewer-theme="night"] .contents-view ol li a:hover {
-            background: #5b5852;
-            color: #fff;
-          }
-        }
-        [data-viewer-theme="night"] .contents-view ol li a.active {
-          background: #dadada;
-          color: #111;
-        }
-        [data-viewer-theme="night"] .contents-view ol li a.active:hover {
-          background: #111;
-          color: #dadada;
-        }
-        [data-viewer-theme="night"] .contents-view ol li span {
-          color: #dadada;
-          border-bottom: 1px solid #444444;
-        }
-
-        .settings-view {
-          overflow: hidden;
-          -webkit-user-select: none;
-          -moz-user-select: none;
-          -ms-user-select: none;
-          user-select: none;
-        }
-        .settings-view > .settings-menu {
-          background-color: #f9f9f9;
-          border: 0.0625rem solid #dadada;
-          border-radius: 0.375rem;
-          box-shadow: 0 0 4px 1px rgba(218, 218, 218, 0.5);
-          right: 0;
-          margin: 0 0.25rem;
-          padding: 0;
-          position: absolute;
-          min-width: 240px;
-          width: auto;
-          max-height: 90vh;
-          overflow-x: hidden;
-          overflow-y: auto;
-          -webkit-overflow-scrolling: touch;
-          box-sizing: border-box;
-        }
-        .settings-view > .settings-menu button {
-          background-color: #fff;
-          border: 0.0625rem solid #dadada;
-          font-size: 0.75rem;
-          letter-spacing: 0.07rem;
-          padding: 0.5rem 1.5rem;
-          text-transform: uppercase;
-          margin: 0;
-          position: relative;
-        }
-        .settings-view > .settings-menu button svg.checkedIcon {
-          color: #fff;
-          display: none;
-          fill: #fff;
-          height: 0.9rem;
-          width: 0.9rem;
-        }
-        .settings-view > .settings-menu button.active {
-          color: #fff;
-          background-color: #468028;
-        }
-        .settings-view > .settings-menu button.active svg.checkedIcon {
-          display: inline-block;
-          margin: 0;
-          position: absolute;
-          top: 50%;
-          transform: translateY(-55%);
-          right: 0.35rem;
-          vertical-align: text-bottom;
-        }
-        .settings-view > .settings-menu > li {
-          border: 0;
-          background-color: #fff;
-          color: #5b5852;
-          display: block;
-          margin: 0.5rem 0;
-          padding: 0;
-          text-align: center;
-        }
-        .settings-view > .settings-menu .settings-options {
-          padding-left: 0;
-          display: flex;
-        }
-        .settings-view > .settings-menu .settings-options li {
-          border: 0;
-          display: flex;
-          margin: 0;
-          width: auto;
-          flex: 1 1 auto;
-        }
-        .settings-view > .settings-menu .settings-options li button {
-          width: 100%;
-          text-overflow: ellipsis;
-        }
-        .settings-view
-          > .settings-menu
-          .settings-options
-          button.publisher-font {
-          border-top-right-radius: 0;
-          border-bottom-right-radius: 0;
-          border-left: 0;
-          border-right: 0;
-        }
-        .settings-view > .settings-menu .settings-options button.serif-font {
-          border-radius: 0;
-        }
-        .settings-view > .settings-menu .settings-options button.sans-font {
-          border-top-left-radius: 0;
-          border-bottom-left-radius: 0;
-          border-left: 0;
-          border-right: 0;
-        }
-        .settings-view > .settings-menu .settings-options button.decrease {
-          border-top-right-radius: 0;
-          border-bottom-right-radius: 0;
-          border-left: 0;
-        }
-        .settings-view > .settings-menu .settings-options button.increase {
-          border-top-left-radius: 0;
-          border-bottom-left-radius: 0;
-          border-left: 0;
-          border-right: 0;
-        }
-        .settings-view > .settings-menu .settings-options button.day-theme {
-          background-color: #fff;
-          border-top-right-radius: 0;
-          border-bottom-right-radius: 0;
-          border-left: 0;
-          border-right: 0;
-        }
-        .settings-view
-          > .settings-menu
-          .settings-options
-          button.day-theme.active {
-          color: #111;
-        }
-        .settings-view
-          > .settings-menu
-          .settings-options
-          button.day-theme.active
-          svg.checkedIcon {
-          fill: #111;
-        }
-        .settings-view > .settings-menu .settings-options button.sepia-theme {
-          background-color: #f6ecd9;
-          border-radius: 0;
-        }
-        .settings-view
-          > .settings-menu
-          .settings-options
-          button.sepia-theme.active {
-          color: #111;
-        }
-        .settings-view
-          > .settings-menu
-          .settings-options
-          button.sepia-theme.active
-          svg.checkedIcon {
-          fill: #111;
-        }
-        .settings-view > .settings-menu .settings-options button.night-theme {
-          background-color: #111;
-          color: #fff;
-          border-top-left-radius: 0;
-          border-bottom-left-radius: 0;
-          border-left: 0;
-          border-right: 0;
-        }
-        .settings-view
-          > .settings-menu
-          .settings-options
-          button.night-theme.active {
-          color: #fff;
-        }
-        .settings-view
-          > .settings-menu
-          .settings-options
-          button.night-theme.active
-          svg.checkedIcon {
-          fill: #fff;
-        }
-        .settings-view
-          > .settings-menu
-          .settings-options
-          button.columns-paginated-view {
-          border-top-right-radius: 0;
-          border-bottom-right-radius: 0;
-          border-left: 0;
-        }
-        .settings-view
-          > .settings-menu
-          .settings-options
-          button.scrolling-book-view {
-          border-top-left-radius: 0;
-          border-bottom-left-radius: 0;
-          border-left: 0;
-          border-right: 0;
-        }
-        .settings-view > .settings-menu .settings-options .icon {
-          color: #5b5852;
-          fill: #5b5852;
-          height: 1.75rem;
-          width: 1.75rem;
-        }
-        @media (min-width: 30rem) {
-          .settings-view > .settings-menu .settings-options .icon {
-            height: 1.5rem;
-            width: 1.5rem;
-          }
-        }
-
-        [data-viewer-theme="night"] .settings-view > .settings-menu {
-          background-color: #111;
-          border: 0.0625rem solid #5b5852;
-          box-shadow: 0 0 4px 1px rgba(218, 218, 218, 0.5);
-        }
-        [data-viewer-theme="night"] .settings-view > .settings-menu button {
-          background-color: #2b2b2b;
-          border: 0.0625rem solid #8e8e8e;
-        }
-        [data-viewer-theme="night"]
-          .settings-view
-          > .settings-menu
-          button
-          svg.checkedIcon {
-          color: #111;
-          fill: #111;
-        }
-        [data-viewer-theme="night"]
-          .settings-view
-          > .settings-menu
-          button.active {
-          color: #fff;
-          background-color: #468028;
-        }
-        [data-viewer-theme="night"] .settings-view > .settings-menu > li {
-          border: 0;
-          background-color: #2b2b2b;
-          color: #fff;
-        }
-        [data-viewer-theme="night"]
-          .settings-view
-          > .settings-menu
-          .settings-options
-          button.day-theme,
-        [data-viewer-theme="night"]
-          .settings-view
-          > .settings-menu
-          .settings-options
-          button.sepia-theme {
-          color: #111;
-        }
-        [data-viewer-theme="night"]
-          .settings-view
-          > .settings-menu
-          .settings-options
-          button.night-theme.active {
-          background-color: #111;
-        }
-        [data-viewer-theme="night"]
-          .settings-view
-          > .settings-menu
-          .settings-options
-          .icon {
-          color: #fff;
-          fill: #fff;
-        }
-
-        /*# sourceMappingURL=main.css.map */
-      `}</style>
-    </>
-  );
+  return <WebpubViewer />;
 };
 
 export default BookPage;
